requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
pandas
google-generativeai
dateparser
dotenv